requirements.txt
aiohttp==3.8.1
async-timeout==4.0.0
attrs==19.3.0
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
discord.py==2.0.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
requests==2.21.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9
typing-extensions==3.7.4.1
urllib3==1.24.1
websockets==6.0
wikipedia==1.4.0
yarl==1.3.0
youtube-dl==2019.9.28
python-dotenv
gunicorn==20.1.0
uvicorn==0.18.2
Flask==2.3.2
google-generativeai
sqlalchemy
discord-py-interactions
